third_party/WebKit/LayoutTests/synthetic_gestures/synthetic-pinch-zoom-gesture.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
  body, html {
    width: 100%;
    height: 100%;
    margin: 0;
  }
  div {
    background-color: green;
    width: 100px;
    height: 100px;
    position: absolute;
    left: 50%;
    top: 50%;
    margin-left: -50px;
    margin-top: -50px;
  }
</style>

<script>
  const kScaleEpsilon = 0.0001;
  const kOffsetEpsilon = 1;

  const t = async_test(
      "This tests that gpuBenchmarking.pinchBy is relatively accurate.");

  function performZoom(scale, anchorX, anchorY, speed, callback) {
    chrome.gpuBenchmarking.pinchBy(scale, anchorX, anchorY, callback, speed);
  }

  const centerX = window.innerWidth / 2;
  const centerY = window.innerHeight / 2;

  // TODO(bokan): Add more variations (speed, anchor location, etc.) but it's
  // difficult right now as pinchBy is badly broken on desktops:
  // https://crbug.com/787615.
  const tests = [
    { starting_scale: 1, scale: 3, speed: 1000, msg: "Zooming in quickly has correct " },
    { starting_scale: 3, scale: 0.5, speed: 1000, msg: "Zooming out quickly has correct " },
    { starting_scale: 1, scale: 3, speed: 100, msg: "Zooming in slowly has correct " },
    { starting_scale: 4, scale: 0.5, speed: 100, msg: "Zooming out slowly has correct " },
    { starting_scale: 1, scale: 3, speed: 50000, msg: "Zooming in instantly has correct " },
    { starting_scale: 4, scale: 0.5, speed: 50000, msg: "Zooming out instantly has correct " },
  ];

  function runTest(testCase) {
    return new Promise((resolve, reject) => {
      internals.setPageScaleFactor(testCase.starting_scale);
      internals.setVisualViewportOffset(
          window.innerWidth * (1 - (1 / testCase.starting_scale)) / 2,
          window.innerHeight * (1 - (1 / testCase.starting_scale)) / 2);

      chrome.gpuBenchmarking.pinchBy(testCase.scale, centerX, centerY, () => {
        t.step(() => {
            const expectedScale = testCase.starting_scale * testCase.scale;
            assert_approx_equals(
                window.visualViewport.scale,
                expectedScale,
                kScaleEpsilon,
                testCase.msg + " page scale factor.");
            assert_approx_equals(
                window.visualViewport.offsetLeft,
                window.innerWidth * (1 - (1 / expectedScale)) / 2,
                kOffsetEpsilon,
                testCase.msg + " visual viewport offsetLeft.");
            assert_approx_equals(
                window.visualViewport.offsetTop,
                window.innerHeight * (1 - (1 / expectedScale)) / 2,
                kOffsetEpsilon,
                testCase.msg + " visual viewport offsetTop.");
        });
        resolve();
      }, testCase.speed);
    });
  }

  async function runAllTestCases() {
    for (const testCase of tests) {
      await runTest(testCase);
    }
    t.done();
  }

  addEventListener('load', runAllTestCases);
</script>
